test: address benchmark review feedback
